test: use rstest projects
