+++ package.json
@@ -19,8 +19,8 @@
     "test:benchmark": "cd ./tests && pnpm run test:benchmark",
     "test:e2e": "pnpm run build:examples && cd tests && pnpm run test:e2e",
     "test:ecosystem-ci": "pnpm run test:unit && cross-env ECO_CI=1 pnpm run test:integration && pnpm run test:e2e",
-    "test:integration": "pnpm -r --filter \"./tests\" run test",
-    "test:unit": "pnpm -r --filter \"./packages/**\" run test",
+    "test:integration": "rstest run --project integration",
+    "test:unit": "rstest run --project unit*",
     "testu": "pnpm run test:unit -u && pnpm run test:integration -u",
     "type-check": "pnpm -r run type-check",
     "update:rsbuild": "npx taze minor --include /rsbuild/ -w -r -l",
@@ -44,7 +44,7 @@
     "@biomejs/biome": "^2.2.2",
     "@changesets/cli": "^2.29.6",
     "@rslint/core": "^0.1.12",
-    "@rstest/core": "^0.3.0",
+    "@rstest/core": "^0.3.1",
     "@types/fs-extra": "^11.0.4",
     "@types/node": "^22.18.0",
     "check-dependency-version-consistency": "^5.0.1",
