+++ packages/create-rslib/fragments/tools/rstest-node-js/package.json
@@ -3,6 +3,6 @@
     "test": "rstest"
   },
   "devDependencies": {
-    "@rstest/core": "^0.3.0"
+    "@rstest/core": "^0.3.1"
   }
 }
