+++ packages/create-rslib/fragments/tools/rstest-react-js/package.json
@@ -3,7 +3,7 @@
     "test": "rstest"
   },
   "devDependencies": {
-    "@rstest/core": "^0.3.0",
+    "@rstest/core": "^0.3.1",
     "@testing-library/jest-dom": "^6.8.0",
     "@testing-library/react": "^16.3.0",
     "jsdom": "^26.1.0"
