+++ packages/create-rslib/fragments/tools/rstest-vue-js/package.json
@@ -3,7 +3,7 @@
     "test": "rstest"
   },
   "devDependencies": {
-    "@rstest/core": "^0.3.0",
+    "@rstest/core": "^0.3.1",
     "@testing-library/jest-dom": "^6.8.0",
     "@testing-library/vue": "^8.1.0",
     "@vue/test-utils": "^2.4.6",
