+++ packages/create-rslib/template-[node-dual]-[rstest]-js/package.json
@@ -19,6 +19,6 @@
   },
   "devDependencies": {
     "@rslib/core": "workspace:*",
-    "@rstest/core": "^0.3.0"
+    "@rstest/core": "^0.3.1"
   }
 }
