+++ packages/create-rslib/template-[node-dual]-[rstest]-ts/package.json
@@ -21,7 +21,7 @@
   },
   "devDependencies": {
     "@rslib/core": "workspace:*",
-    "@rstest/core": "^0.3.0",
+    "@rstest/core": "^0.3.1",
     "@types/node": "^22.18.0",
     "typescript": "^5.9.2"
   }
