+++ packages/create-rslib/template-[node-esm]-[rstest]-js/package.json
@@ -17,6 +17,6 @@
   },
   "devDependencies": {
     "@rslib/core": "workspace:*",
-    "@rstest/core": "^0.3.0"
+    "@rstest/core": "^0.3.1"
   }
 }
