+++ packages/create-rslib/template-[react]-[rstest,storybook]-ts/package.json
@@ -23,7 +23,7 @@
     "@rsbuild/core": "~1.5.1",
     "@rsbuild/plugin-react": "^1.3.5",
     "@rslib/core": "workspace:*",
-    "@rstest/core": "^0.3.0",
+    "@rstest/core": "^0.3.1",
     "@storybook/addon-docs": "^9.1.3",
     "@storybook/addon-essentials": "^9.0.0-alpha.12",
     "@storybook/addon-interactions": "^9.0.0-alpha.10",
