+++ packages/create-rslib/template-[react]-[rstest]-ts/package.json
@@ -20,7 +20,7 @@
   "devDependencies": {
     "@rsbuild/plugin-react": "^1.3.5",
     "@rslib/core": "workspace:*",
-    "@rstest/core": "^0.3.0",
+    "@rstest/core": "^0.3.1",
     "@testing-library/jest-dom": "^6.8.0",
     "@testing-library/react": "^16.3.0",
     "@types/react": "^19.1.12",
