+++ packages/create-rslib/template-[vue]-[rstest,storybook]-js/package.json
@@ -22,7 +22,7 @@
   "devDependencies": {
     "@rsbuild/core": "~1.5.1",
     "@rslib/core": "workspace:*",
-    "@rstest/core": "^0.3.0",
+    "@rstest/core": "^0.3.1",
     "@storybook/addon-docs": "^9.1.3",
     "@storybook/addon-essentials": "^9.0.0-alpha.12",
     "@storybook/addon-interactions": "^9.0.0-alpha.10",
