+++ packages/create-rslib/template-[vue]-[rstest]-js/package.json
@@ -19,7 +19,7 @@
   },
   "devDependencies": {
     "@rslib/core": "workspace:*",
-    "@rstest/core": "^0.3.0",
+    "@rstest/core": "^0.3.1",
     "@testing-library/jest-dom": "^6.8.0",
     "@testing-library/vue": "^8.1.0",
     "@vue/test-utils": "^2.4.6",
