+++ packages/plugin-dts/rstest.config.ts
@@ -1,5 +1,5 @@
 import { defineConfig } from '@rstest/core';
-import { shared } from '../../rstest.workspace';
+import { shared } from '../../rstest.config';
 
 // Disable color in test
 process.env.NO_COLOR = '1';
