+++ tests/rstest.config.ts
@@ -1,6 +1,6 @@
 // import codspeedPlugin from '@codspeed/vitest-plugin';
 import { defineConfig } from '@rstest/core';
-import { shared } from '../rstest.workspace';
+import { shared } from '../rstest.config';
 
 export default defineConfig({
   ...shared,
